edge_service/requirements.txt
# Core web framework and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Rate limiting
slowapi==0.1.9

# Computer vision and image processing
opencv-python==4.8.1.78
numpy==1.24.3

# MQTT communication
paho-mqtt==1.6.1

# Kafka for hard example publishing
kafka-python==2.0.2
confluent-kafka==2.3.0

# Monitoring and metrics
prometheus-client>=0.16.0

# ONNX Runtime for model optimization and quantization
onnxruntime>=1.16.0
onnx>=1.14.0

# System monitoring for performance metrics
psutil>=5.9.0

# Structured logging
python-json-logger==2.0.7
structlog==23.2.0

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
PyJWT==2.8.0

# HTTP client for health checks and OTA updates
requests==2.31.0

# Additional utilities
python-multipart==0.0.6
tqdm==4.66.1

# Observability dependencies
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp==1.21.0

# HashiCorp Consul integration (if needed)
python-consul==1.1.0

# Core dependencies for edge service
pathlib2==2.3.7
Pillow==10.0.1

# Security auditing
pip-audit==2.6.3

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
testcontainers[compose]==3.7.1
